[WOW64] Added non-functional PE build
