+++ src/os/emit_signal_wine.c
@@ -1,0 +1,22 @@
+#include "x64emu.h"
+#include "custommem.h"
+
+void EmitSignal(x64emu_t* emu, int sig, void* addr, int code)
+{
+    // FIXME
+}
+
+void CheckExec(x64emu_t* emu, uintptr_t addr)
+{
+    // FIXME
+}
+
+void EmitInterruption(x64emu_t* emu, int num, void* addr)
+{
+    // FIXME
+}
+
+void EmitDiv0(x64emu_t* emu, void* addr, int code)
+{
+    // FIXME
+}
